Create quiz2.js
Write a function that accepts two string parameters. The first parameter will be a string of characters, and the second parameter will be the same string of characters, but they’ll be in a different order and have one extra character. The function should return that extra character.
For example, if the first parameter is "eueiieo" and the second is "iieoedue," then the function should return "d."